Revert "Force cleaning of sources on spark-x modules"
This reverts commit c9d13856e49b8708a998eb3e9ceac695369c64f0 [formerly 2ce822862427d4eb9812a38af57cdbeb77a179f0 [formerly cfbeb3b921e8a10d829f142f674cf4f5129e1ccc]].

This erased the jar created in a normal 'mvn install' run.
This revert will also create issues in ./buildmultiplescalaversions.sh
runs, where the non-cleaning of tasks will make a spark-2 build pick up a
spark-1 artefact.

The fix would be to 'mvn clean' between changes of versions,
but then we run into #2873

Hotfixing this for now, as ./buildmultiplescalaversions.sh runs
are less frequent.


Former-commit-id: b9bac7f18eeda80ee3dac7effc0f9568a799740e [formerly f21107e3f4fdd4221b8c72216cdd22231bd3cb14]
Former-commit-id: 17e1575d85a5f87361f0cd759fa9847f528790dc